cache: refactor moving from bbgo to pkg/cache
