+++ pkg/cache/home.go
@@ -1,4 +1,4 @@
-package bbgo
+package cache
 
 import (
 	"os"
